add more casts for malloc

+++ src/autlib5.c
@@ -729,7 +729,7 @@ fiho(iap_type *iap, rap_type *rap, int64 ndim, double *par, int64 *icp,
     (void)ndim;
     (void)rap;
 
-    fj = malloc(sizeof(*fj)*(iap->ndim));
+    fj = malloc(sizeof(*fj)*(usize)(iap->ndim));
     /* Generates the integral conditions for homoclinic orbits. */
 
     dint_dim1 = nint;
@@ -798,7 +798,7 @@ inho(iap_type *iap, int64 *icp, double *par) {
 
     blhmp_1.ipsi = malloc(sizeof(*(blhmp_1.ipsi))*NPARX);
     blhmp_1.ifixed = malloc(sizeof(*(blhmp_1.ifixed))*NPARX);
-    blhmp_1.irev = malloc(sizeof(*(blhmp_1.irev))*(iap->ndim));
+    blhmp_1.irev = malloc(sizeof(*(blhmp_1.irev))*(usize)(iap->ndim));
 
     blhme_1.ieigc = malloc(sizeof(*(blhme_1.ieigc))*2);
 
@@ -847,19 +847,18 @@ inho(iap_type *iap, int64 *icp, double *par) {
     /* Allocate memory for global structures.  We didn't know the
        size for these until ndim was computed. */
 
-    blhmu_1.pu0 = malloc(sizeof(*(blhmu_1.pu0))*(ndim));
-    blhmu_1.pu1 = malloc(sizeof(*(blhmu_1.pu1))*(ndim));
+    blhmu_1.pu0 = malloc(sizeof(*(blhmu_1.pu0))*(usize)ndim);
+    blhmu_1.pu1 = malloc(sizeof(*(blhmu_1.pu1))*(usize)ndim);
 
-    bleig_1.rr = malloc(sizeof(*(bleig_1.rr))*(ndim));
-    bleig_1.ri = malloc(sizeof(*(bleig_1.ri))*(ndim));
-    bleig_1.v = malloc(sizeof(*(bleig_1.v))*(ndim)*(ndim));
-    bleig_1.vt = malloc(sizeof(*(bleig_1.vt))*(ndim)*(ndim));
-    bleig_1.xequib = malloc(sizeof(*(bleig_1.xequib))*(ndim));
+    bleig_1.rr = malloc(sizeof(*(bleig_1.rr))*(usize)ndim);
+    bleig_1.ri = malloc(sizeof(*(bleig_1.ri))*(usize)ndim);
+    bleig_1.v = malloc(sizeof(*(bleig_1.v))*(usize)(ndim*ndim));
+    bleig_1.vt = malloc(sizeof(*(bleig_1.vt))*(usize)(ndim*ndim));
+    bleig_1.xequib = malloc(sizeof(*(bleig_1.xequib))*(usize)ndim);
 
-    blhme_1.vrprev = malloc(sizeof(*(blhme_1.vrprev))*2 * (ndim)*(ndim));
+    blhme_1.vrprev = malloc(sizeof(*(blhme_1.vrprev))*(usize)(2*ndim*ndim));
 
-    beyn_1.cprev = malloc(sizeof(*(beyn_1.cprev))*2 * 2*(ndim)*(ndim));
-
+    beyn_1.cprev = malloc(sizeof(*(beyn_1.cprev))*(usize)(2*2*ndim*ndim));
     nfree = blhom_1.nfixed + 2 - blhom_1.nrev + nint + nbc;
     bcnn_1.nbcn = nbc;
 
@@ -1059,7 +1058,7 @@ stpnho(iap_type *iap, rap_type *rap, double *par, int64 *icp, int64 *ntsr,
     (void)rldot;
     (void)rap;
 
-    u = malloc(sizeof(*u)*(iap->ndim));
+    u = malloc(sizeof(*u)*(usize)(iap->ndim));
     /* Generates a starting point for the continuation of a branch of */
     /* of solutions to general boundary value problems by calling the user */
     /* supplied subroutine STPNT where an analytical solution is given. */
@@ -1132,11 +1131,11 @@ stpho(iap_type *iap, int64 *icp, double *u, double *par, double *t) {
     double *ri;
     double *rr, *vr, *vt, *xequib;
 
-    ri = malloc(sizeof(*ri)*(iap->ndm));
-    rr = malloc(sizeof(*rr)*(iap->ndm));
-    vr = malloc(sizeof(*vr)*(iap->ndm)*(iap->ndm));
-    vt = malloc(sizeof(*vt)*(iap->ndm)*(iap->ndm));
-    xequib = malloc(sizeof(*xequib)*(iap->ndm));
+    ri = malloc(sizeof(*ri)*(usize)(iap->ndm));
+    rr = malloc(sizeof(*rr)*(usize)(iap->ndm));
+    vr = malloc(sizeof(*vr)*(usize)((iap->ndm)*(iap->ndm)));
+    vt = malloc(sizeof(*vt)*(usize)((iap->ndm)*(iap->ndm)));
+    xequib = malloc(sizeof(*xequib)*(usize)(iap->ndm));
 
     /* Generates a starting point for homoclinic continuation */
     /* If ISTART=2 it calls STPNHO. */
@@ -1340,8 +1339,8 @@ psiho(iap_type *iap, int64 is, double *rr, double *ri, double *v, double *vt,
     int64 ndm;
     double dum1, dum2;
 
-    f0 = malloc(sizeof(*(f0))*(iap->ndm));
-    f1 = malloc(sizeof(*(f1))*(iap->ndm));
+    f0 = malloc(sizeof(*(f0))*(usize)(iap->ndm));
+    f1 = malloc(sizeof(*(f1))*(usize)(iap->ndm));
 
     /* The conditions for degenerate homoclinic orbits are given by PSI(IS)=0.
      */
@@ -1599,9 +1598,9 @@ eighi(int64 isign, int64 itrans, double *rr, double *ri, double *vret,
     double *dfdp, *dfdu;
     double *zz;
 
-    dfdp = malloc(sizeof(*dfdp)*(*ndm)*NPARX);
-    dfdu = malloc(sizeof(*dfdu)*(*ndm)*(*ndm));
-    zz = malloc(sizeof(*zz)*(*ndm)*(*ndm));
+    dfdp = malloc(sizeof(*dfdp)*(usize)((*ndm)*NPARX));
+    dfdu = malloc(sizeof(*dfdu)*(usize)((*ndm)*(*ndm)));
+    zz = malloc(sizeof(*zz)*(usize)((*ndm)*(*ndm)));
 
     eigho(&isign, &itrans, rr, ri, vret, xequib, icp, par, ndm, dfdu, dfdp, zz);
